# documents/api/API.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "ZaloPay "
  description: BigAssignment
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/oispmdp/BigAssignment/1.0
...
paths:
...
  /payment/api/v1/systems/getAllWallets:
    get:
      tags:
        - Payment Service
      description: >-
        System get all wallet
      parameters:
        - in: header
          name: Authorization
          description: Bearer system token
          schema:
            type: string
          required: true
          example: Bearer kasfdkjajfjaf.kjadsfkjakjfdhjagae.rkjsdkjakjdsf
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSystemResponse"
        "401":
          description: Unauthorize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
  /payment/api/v1/systems/createTransferFromUser:
    post:
      tags:
        - Payment Service
      description: >-
        Create transfer from user
      parameters:
        - in: header
          name: Authorization
          description: Bearer system token
          schema:
            type: string
          required: true
          example: Bearer kasfdkjajfjaf.kjadsfkjakjfdhjagae.rkjsdkjakjdsf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemCreateTransferFromUserRequest"
        description: User login to the app
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: "Message"
                  payload:
                    type: object
                    properties:
                      amount:
                        type: number
                        example: 30000
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
        "401":
          description: Unauthorize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"
        "404":
          description: "Not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
  /payment/api/v1/systems/createTransferToUser:
    post:
      tags:
        - Payment Service
      description: >-
        Create transfer to user
      parameters:
        - in: header
          name: Authorization
          description: Bearer system token
          schema:
            type: string
          required: true
          example: Bearer kasfdkjajfjaf.kjadsfkjakjfdhjagae.rkjsdkjakjdsf
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemCreateTransferToUserRequest"
        description: Create transfer to user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: "Transfer to user successfully"
                  payload:
                    type: string
                    example: "nothing"
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error400"
        "401":
          description: Unauthorize
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error401"

  ### Lucky Money

  # Create Lucky Money from user to a group chat
...
components:
  schemas:
...
    SystemCreateTransferFromUserRequest:
      type: object
      properties:
        userId:
          type: string
          example: uuid-123
        walletId:
          type: number
          example: 1234
        softToken:
          type: string
          example: adsfadsfafd.fafafafafsfdadsf.dfasdfafdasf
        message:
          type: string
          example: Happy new year
        amount: 
          type: number
          example: 30000
...
    WalletInfo:
      type: object
      required:
        - walletId
        - balance
      properties:
        balance:
          type: number
          example: 200000
        walletId:
          type: number
          example: 123
    WalletSystemResponse:
      type: object
      required:
        - success
        - code
        - message
        - payload
      properties:
        success:
          type: boolean
          default: true
          example: true
        code:
          type: integer
          default: 200
          example: 200
        message:
          type: string
          default: "Message"
        payload:
          type: array
          items:
            $ref: "#/components/schemas/WalletInfo"
...
    SystemCreateTransferToUserRequest:
      type: object
      required:
        - walletId
        - receiverId
        - amount
        - message
      properties:
        walletId:
          type: number
          example: 1
        receiverId:
          type: number
          example: 1
        amount:
          type: number
          example: 100000
        message:
          type: string
          example: Goodbye!
...
    Error400:
      type: object
      required:
        - success
        - code
        - message
        - payload
      properties:
        success:
          type: boolean
          default: false
          example: false
        code:
          type: integer
          default: 400
          example: 400
        message:
          type: string
          default: "Bad Request"
        payload:
          type: object
          default: {}
          example: {}
    # =====================
...
    Error401:
      type: object
      required:
        - success
        - code
        - message
        - payload
      properties:
        success:
          type: boolean
          default: false
          example: false
        code:
          type: integer
          default: 401
          example: 401
        message:
          type: string
          default: "Unauthorized"
        payload:
          type: object
          default: {}
          example: {}
    # =====================
...